refactor(docs): split content library from app config wiring
